requirements.txt
pyswisseph
RPi.GPIO
pyserial
pytz
